refactor(auth): move refresh-token flow into freeq-oauth
Extract the broker's refresh_access_token, RefreshError, and is_invalid_grant
into freeq_oauth::flow, re-signatured to primitives (injected reqwest::Client,
token_endpoint, client_id, dpop_key, refresh_token, current_nonce) returning a
RefreshedTokens struct. This is the durable-session logic the embedded mount
will reuse in Phase 3.

The broker keeps a thin refresh_access_token(config, record) adapter that
builds client_id from config and injects its bounded HTTP client. Behavior is
unchanged — the 23 characterization tests (single-use rotation, concurrent
refresh-lock serialization, 401-vs-502 classification) pass. The three
refresh-classification unit tests moved with the code into freeq-oauth.

freeq-oauth: 8 unit tests. Broker: 23 characterization tests green.
